+++ tests/ShopSharp.Core.Domain.UnitTests/AggregateTests/Fakes/FakeDomainEvent.cs
@@ -1,4 +1,4 @@
-﻿using ShopSharp.Core.Domain.Aggregates;
+﻿using ShopSharp.Core.Domain.Events;
 
 namespace ShopSharp.Core.Domain.AggregateTests.Fakes;